# Fix isequal(x::Tuple{Ptr{UInt8}, Int}, y::String)