We cannot rely on hashes being equal to conclude that strings are equal since hash collisions can happen.

I considered using `x[1] == pointer(y) && x[2] == sizeof(y) && return true` at the top of the function as in Base, to take fast path when strings point to the same memory, but I guess in practice it will almost never happen that an existing string points inside a buffer, right?