@quinnj 2019-01-29
Can you show what the performance impact is here?