@nalimilan 2019-01-29
It's much more efficient, since computing hashes is costly and cannot return early when sizes differ:
```julia
using Parsers, BenchmarkTools

y = "fskjqmlqkjkjkrze"
y2 = "fskjqmlqkjkjkrze2"
x = (pointer(y), sizeof(y))
@btime isequal($x, $y);
@btime isequal($x, $y2);

# master
julia> @btime isequal($x, $y2);
  19.075 ns (0 allocations: 0 bytes)

julia> @btime isequal($x, $y2);
  19.772 ns (0 allocations: 0 bytes)

# This PR
julia> @btime isequal($x, $y);
  4.516 ns (0 allocations: 0 bytes)

julia> @btime isequal($x, $y2);
  2.169 ns (0 allocations: 0 bytes)
```